requirements.txt
Django==1.7
Fabric==1.9.1
Markdown==2.4.1
astroid==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-tools==0.5.2
django-model-utils==2.2
django-mptt==0.6.1
django-toolbelt==0.0.1
ecdsa==0.11
gunicorn==19.1.1
logilab-common==0.62.1
paramiko==1.14.1
pep8==1.5.7
psycopg2==2.5.4
pycrypto==2.6.1
pylint==1.3.1
static3==0.5.1
wsgiref==0.1.2
